backend/requirements.txt
Flask>=3.1,<4
flask-cors>=5.0,<6
python-dotenv>=1.0,<2
pymongo>=4.10,<5
requests>=2.32,<3
beautifulsoup4>=4.12,<5
selenium>=4.34,<5
webdriver-manager>=4.0,<5
